Update forwarder.cpp
Implement onFibUpdate. Its main function is to check whether there are any routes in the Fib table that have changed from unreachable to reachable, then mark the global route as updated and call clearFloodPackets. In the constructor of Forwarder, register the update listener of the Fib table to ensure that the onFibUpdate method is triggered whenever the Fib table changes. The clearFloodPackets function is designed to clear the flood packets in the network when the route update is completed.